requirements.txt
click==8.1.8
dlib==19.24.6
face-recognition==1.3.0
face-recognition-models==0.3.0
imutils==0.5.4
joblib==1.4.2
numpy==1.26.4
opencv-python==4.10.0.84
pillow==11.1.0
pyzxing==1.0.2
qrcode==8.0
